backend-unified/requirements.txt
# MarketingFunnelMaster Unified Backend - Requirements
# Version: 1.0.0
# Updated: 2025-07-03

# Core FastAPI Stack
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async (migration buffer)
alembic==1.13.1  # Database migrations

# Supabase Integration
supabase==2.0.2

# Caching & Session Storage
redis==5.0.1
aioredis==2.0.1

# Background Tasks
celery==5.3.4

# Real-time Communication
websockets==12.0

# AI & Machine Learning (Existing Integration Enhanced)
openai==1.3.7  # GPT-4 Integration
anthropic==0.7.8  # Claude Integration
google-cloud-aiplatform==1.38.1  # Gemini Integration

# Vector Database & Knowledge Graph
chromadb==0.4.18
networkx==3.2.1

# Natural Language Processing
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2

# HTTP Client & File Operations
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# Web Scraping & Browser Automation
beautifulsoup4==4.12.2
selenium==4.16.0
playwright==1.40.0

# Image Processing
pillow==10.1.0

# Data Processing & Analysis
pandas==2.1.4
numpy==1.24.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Environment & Configuration
python-dotenv==1.0.0

# Date & Time Utilities
python-dateutil==2.8.2

# Task Scheduling
apscheduler==3.10.4

# Email Services
sendgrid==6.10.0
python-mailgun2==1.2.1

# Cloud Infrastructure APIs
hcloud==1.33.1  # Hetzner Cloud API
cloudflare==2.11.1

# Development & Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing
black==23.11.0  # Code formatting
isort==5.12.0  # Import sorting
flake8==6.1.0  # Linting
mypy==1.7.1  # Type checking

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Performance Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Rate Limiting
slowapi==0.1.9

# File Format Support
openpyxl==3.1.2  # Excel files
PyPDF2==3.0.1  # PDF processing

# Async Task Management
asyncio-mqtt==0.16.1
aiokafka==0.8.11

# Additional Utilities
click==8.1.7  # CLI tools
rich==13.7.0  # Rich terminal output
typer==0.9.0  # CLI framework

# Migration Support (for Express.js → FastAPI transition)
pyjwt==2.8.0
bcrypt==4.1.2

# WebSocket Extensions
python-socketio==5.10.0

# Graph Database (optional for advanced knowledge graphs)
neo4j==5.14.1

# Time Series Database (for analytics)
influxdb-client==1.38.0

# Message Queue (Redis-based)
rq==1.15.1

# API Documentation Enhancement
fastapi-users==12.1.2
fastapi-limiter==0.1.5

# File Upload & Storage
python-magic==0.4.27
boto3==1.34.0  # AWS S3 compatibility

# Internationalization
babel==2.14.0

# Health Checks
health-check==1.3.3

# Environment Detection
distro==1.8.0

# JSON Schema Validation
jsonschema==4.20.0

# Template Engine (for website generation)
jinja2==3.1.2

# CSS/JS Minification
cssmin==0.2.0
jsmin==3.0.1

# HTML Parsing & Generation
lxml==4.9.3
markdown==3.5.1

# Cryptography Enhancements
cryptography==41.0.7

# Development Server Enhancements
watchfiles==0.21.0

# API Testing & Documentation
tavern==2.6.1